next error fixing in uvc_video

+++ Linux_uvc_driver/uvc_3.3.8/uvc_video.c
@@ -679,11 +679,11 @@ void uvc_video_clock_update(struct uvc_streaming *stream,
 	}
 
 	uvc_trace(UVC_TRACE_CLOCK, "%s: SOF %u.%06llu y %llu ts %lld.%06lld "
-		  "buf ts %lu.%06lu (x1 %u/%u/%u x2 %u/%u/%u y1 %u y2 %u)\n",
+		  "buf ts %lld.%06lld (x1 %u/%u/%u x2 %u/%u/%u y1 %u y2 %u)\n",
 		  stream->dev->name,
 		  sof >> 16, div_u64(((u64)sof & 0xffff) * 1000000LLU, 65536),
 		  y, (long long)ts.tv_sec, (long long)(ts.tv_nsec / NSEC_PER_USEC),
-		  v4l2_buf->timestamp.tv_sec, v4l2_buf->timestamp.tv_usec,
+		  (long long)v4l2_buf->timestamp.tv_sec, (long long)v4l2_buf->timestamp.tv_usec,
 		  x1, first->host_sof, first->dev_sof,
 		  x2, last->host_sof, last->dev_sof, y1, y2);
